Sort entries after determining fill-in
Not needed for correcteness, but improves performance of factorization.
Generation of nonzeros using the (not-postordered) elimination tree
already sorts entries that are in the same row next to eachother.
However, entries within a row are in general not sorted by column. So
just call std::sort for now. Happens in symbolic analysis phase so not
worrying about making it real fast for now.