Refactor passing of constexpr data
Pass sparsity and permutation as object to functions instead of only as
type. That allows introducing a common Sparsity class to reduce the
number of different types needed. For example user code can use this
Sparsity class to simplify passing their sparsity.
If values are needed at compile-time within a constexpr function itself
(for example for constructing an array of certain size), using non-type
template parameters for passing now.