Fix permutation for single factorization
A direct user of the Factorization class for factorizing a single matrix
(as opposed to repeating block tridiagonals) would expect their input
sparsity to be permuted if they also pass a permutation. That was not
the case, but is fixed now.

The problem is that the repeating block tridiagonal factorization cannot
use that single factorization class directly anymore, since the
sparsity of the diagonal block is already permuted to compute the
stable repeating sparsity. Applying Factorization directly would mean
that permutation would be applied again. Fixing that by unpermuting the
stable repeating filled-in sparsity and then passing it on.