More consistent handling of row_offset
Since we are iterating with the row_index offset by row_offset, we
should also call the given function with that row_offset. That means
that the entries of one block down in the repeated structure will be
iterated on, which makes it a bit more awkward there. But this makes the
function more useful for the general case.