Simplify, clarify and document backward solve
For the backward substitution, there is actually no advantage to taking
the diagonal and the subdiagonal block at the same time, since there is
no solution entry that can be reused as in the forward substitution. So
splitting this up makes the function simpler and more versatile. Also
improve parameter names and add documentation.