Fix factorization making up unnecessary nonzeros
When applying the start block to the first tridiagonal and outer blocks,
only the start permutation was considered, not the tridiagonal and outer
ones. Since the 3x3 fill-in already computed the fill-in within
tridiagonal and outer itself, unnecessary fill-in was generated due to
the wrong permutation. To fix that, already use the tridiagonal and
outer permutation there and then do not permute in the following steps.